Add files via upload
This is a Machine Learning model which predicted the user's preferences on the online Music store app.
During the training model, we have used a dataset in CSV format, the dataset includes attributes such as age ranges and gender of users. The backend development suggests a user sign up on the App by inserting a password, specifying the year and gender. the model has tried to learn these expected inputs and suggest music according to the form submitted.

The App will suggest a user a song according to his age and preferences. the music kind was pop, Jazz and Hipop. The model predicted that the users of between 20-25 years will mostly listed to HIP song and 30-40 will likely to listen to Jazz.